+++ NEWS.rst
@@ -6,8 +6,9 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
-
+* Minor: Added decoder functions; ``is_symbol_uncoded``, ``is_symbol_missing``,
+  and ``is_symbol_partially_decoded``.
+* Patch: Updated documentation
 10.0.0
 ------
 * Major: Depend on kodo-core, kodo-rlnc and kodo-fulcrum instead of kodo